Code clean-up
